Refactored everything to use rust_decimal and NUMERIC(15,4) in the database for currency.

+++ backend/src/util/locale_context.rs
@@ -12,6 +12,7 @@ use chrono::{
     NaiveDate,
 };
 use fluent::FluentArgs;
+use rust_decimal::Decimal;
 use shared_core::i18n::format_date_icu;
 
 /// A lightweight, stateless context helper to match frontend ergonomics on the backend.
@@ -41,12 +42,12 @@ impl<'a> LocaleContext<'a> {
     }
 
     /// Standard currency formatting ("1,234.56")
-    pub(crate) fn format_money(&self, amount_cents: i64) -> String {
+    pub(crate) fn format_money(&self, amount_cents: Decimal) -> String {
         shared_core::i18n::format_currency_icu(amount_cents, Some(self.locale)) //
     }
 
     /// Typst-safe currency formatting ("−1,234.56")
-    pub(crate) fn format_money_typ(&self, amount_cents: i64) -> String {
+    pub(crate) fn format_money_typ(&self, amount_cents: Decimal) -> String {
         shared_core::i18n::format_currency_icu_typ(amount_cents, Some(self.locale))
         //
     }
