+++ frontend/src/payables/components/vendor_invoice_drawer/item_edit_card.rs
@@ -5,7 +5,7 @@
  * called LICENSE at the top level of the KelpieBooks source tree
  *  (online at: https://github.com/shartrec/kelpiebooks/LICENSE ).
  */
-
+use rust_decimal::Decimal;
 use shared_core::{
     ledger::models::account::Account,
     payables::models::vendor_invoice_item::VendorInvoiceItem,
@@ -56,7 +56,7 @@ pub fn item_edit_card(props: &ItemEditCardProps) -> Html {
 
     let on_net_amount_change = {
         let item = item.clone();
-        Callback::from(move |value: i64| {
+        Callback::from(move |value: Decimal| {
             let mut new_item = (*item).clone();
             new_item.net_amount = value;
             new_item.total_amount = new_item.net_amount + new_item.tax_amount;
@@ -66,7 +66,7 @@ pub fn item_edit_card(props: &ItemEditCardProps) -> Html {
 
     let on_tax_amount_change = {
         let item = item.clone();
-        Callback::from(move |value: i64| {
+        Callback::from(move |value: Decimal| {
             let mut new_item = (*item).clone();
             new_item.tax_amount = value;
             new_item.total_amount = new_item.net_amount + new_item.tax_amount;
