+++ shared_core/src/payables/requests/vendor_invoice_payment.rs
@@ -7,6 +7,7 @@
  */
 
 use chrono::NaiveDate;
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
@@ -15,6 +16,6 @@ pub struct CreateVendorInvoicePaymentRequest {
     pub vendor_invoice_id: Uuid,
     pub payment_date: NaiveDate,
     pub bank_account_id: Uuid,
-    pub amount: i64,
+    pub amount: Decimal,
     pub reference: Option<String>,
 }
