+++ shared_core/src/sales/dtos/item_payload.rs
@@ -5,6 +5,7 @@
  * called LICENSE at the top level of the KelpieBooks source tree
  *  (online at: https://github.com/shartrec/kelpiebooks/LICENSE ).
  */
+use rust_decimal::Decimal;
 use serde::Deserialize;
 use uuid::Uuid;
 use crate::sales::models::item::ItemType;
@@ -16,7 +17,7 @@ pub struct ItemPayload {
     pub description: Option<String>,
     pub item_type: ItemType,
     pub uom_id: Uuid,
-    pub unit_price: i64,                    // 💡 Scaled to 4 decimal places internally
+    pub unit_price: Decimal,                    // 💡 Scaled to 4 decimal places internally
     pub income_account_id: Uuid,
     pub tax_category_id: Option<Uuid>,
     pub is_active: bool,
